Norme

+++ philo/time.c
@@ -6,13 +6,13 @@
 /*   By: igchurru <igchurru@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/28 15:12:22 by igchurru          #+#    #+#             */
-/*   Updated: 2025/01/31 12:07:13 by igchurru         ###   ########.fr       */
+/*   Updated: 2025/02/03 12:49:57 by igchurru         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "philo.h"
 
-ssize_t	ft_get_current_time(void)
+size_t	ft_get_current_time(void)
 {
 	struct timeval	time;
 
@@ -21,7 +21,7 @@ ssize_t	ft_get_current_time(void)
 		write(2, "gettimeofday() error\n", 21);
 		return (-1);
 	}
-	return ((time.tv_sec * 1000) + (time.tv_usec / 1000));
+	return ((size_t)(time.tv_sec * 1000) + (time.tv_usec / 1000));
 }
 
 int	ft_usleep(size_t milliseconds)
